pyproject.toml
[tool.poetry]
name = "tom-nonlocalizedevents"
version = "0.5.3"
description = "Reusable TOMToolkit app to support gravitational wave superevent and other nonlocalized event EM follow-up observations."
authors = ["TOM Toolkit Project <tomtoolkit@lco.global>", "David Collom <dcollom@lco.global>", "Lindy Lindstrom <llindstrom@lco.global>", "Jonathan Nation <jnation@lco.global>"]
license = "GPL-3.0-only"
readme = "README.md"
repository = "https://github.com/TOMToolkit/tom_nonlocalizedevents"
keywords = ["tomtoolkit", "astronomy", "astrophysics", "cosmology", "science"]
classifiers = [
    "Environment :: Web Environment",
    "Framework :: Django",
    "Framework :: Django :: 3.1",
    "Intended Audience :: Developers",
    "Intended Audience :: Science/Research",
    "License :: OSI Approved :: GNU General Public License v3 or later (GPLv3+)",
    "Operating System :: OS Independent",
    "Programming Language :: Python",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3 :: Only",
    "Programming Language :: Python :: 3.8",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Topic :: Scientific/Engineering :: Astronomy",
    "Topic :: Scientific/Engineering :: Physics"
]

packages = [
    { include = "tom_nonlocalizedevents" }
]

[tool.setuptools_scm]

[tool.poetry.dependencies]
python = ">=3.8,<3.12"
gracedb-sdk = "^0.1"
python-dateutil = "^2.8"
django-webpack-loader = "1.6.0"
tomtoolkit = "^2.10"
voevent-parse = "^1.0"
django-filter = ">=21.0 <22.0"

[tool.poetry.dev-dependencies]
coverage = "^6.3.2"
factory_boy = "^3.1.0"
pytest = "^3.0.3"

[build-system]
requires = ["poetry-core>=1.0.0", "setuptools>=45", "setuptools_scm[toml]>=6.2"]
build-backend = "poetry.core.masonry.api"
